Make storage ID's optional in @JSimpleClass, etc. annotations.
Add new StorageIdGenerator interface for auto-generating storage ID's,
plus a DefaultStorageIdGenerator standard implementation.

In DefaultStorageIdGenerator hash the item type and name to get a random value
in the range 0x1fb-0x100fb, which is the range that encodes in three bytes.
This provides for a collision space of 65,280 storage IDs.

Add new JSimpleDBFactory class for configuring a new JSimpleDB, allowing
custom StorageIdGenerator's to be used if desired.