Fix an overly pessimistic Raft follower corner case.
If a follower receives a log entry with an index less that the follower's
last applied index, then Raft guarantees that the corresponding term must
match. So the follower should reply with success instead of failure.

This change allows us to also remove code in leaders that tries to avoid
sending such log entries in the first place.