Fix possible data corruption race condition in SimpleKVDatabase.
The methods getAtLeast() and getAtMost() were finding the next key prior and
then acquiring a read lock on the key range scanned.  The problem is that when
acquiring a read lock, the thread can sleep with the object monitor released.

If another thread has mutated the scanned range, and therefore aquired a write
lock, but not yet committed, the first thread will block until it does commit.
If the second thread commits while the first thread sleeps, the second thread's
mutations will then be visible to the first thread after it wakes up, whereas
previously they were not, violating isolation.

This fixs locks to the end of the key range, which is much more pessimistic than
before. A more sophisticated algorithm might rescan the key range after acquring
the lock to verify nothing has changed, etc.