Add workaround for Woodstox issue #17.
When woodstox is on the classpath, it takes over as the default
XMLOutputFactory implementation. However, it's behavior is not
entirely compatible with the usual default; see this issue:

    https://github.com/FasterXML/woodstox/issues/17

This happens when the Maven plug-in is used, because woodstox
is pulled in by Maven. So add a simple workaround for this problem.