Handle fields declared in multiple incomparable supertypes.
When a JField is declared in two supertypes, but not the model class itself,
and the two supertypes are not assignable to each other, then we end up
attempting to create the same JField twice. Previously this would always
cause an error. Now we allow that this can happen and instead check whether
the two fields are actually equivalent before throwing an error.