Fix a bug in the Raft follower algorithm when probing is enabled.
The bug is in the modification to the follower algorithm in which we require a
successful ping response from a majority of peers before becoming a candidate.

We should not wait for the next election timeout to determine whether we have
received ping responses from a majority of nodes, because this can lead to an
election loop when (a) there are only two nodes in the Raft cluster and (b) we
have the longest log: we will always reject the peer's vote request, yet because
of the extra time delay of one election timeout (due to probing) we never get to
become a candidate in the next term, because the peer always does so first.

The fix for this is to detect a majority of ping responses as soon as the last
required ping response is received, instead of waiting for the next election
timeout, so that the additional time due to probing before we become a candidate
is minimal relative to the election timeout.