Add support for guaranteeing that references are always valid.
Currently it's possible to set a reference field to a deleted object, or to
copy an object containing a valid reference to another transaction in which
the referenced object does not exist.

Both of these loopholes allow the counter-intuitive creation of dangling
references.

To control this behavior, add new @JField annotations for regular and snapshot
transactions, and set the default for regular transactions to disallow it.
Make this default incompatible with DeleteAction.NOTHING.

Update the various copy() methods to enforce reference validation, while
continuing to allow batch copies of object graphs having cyclic references.