Treat a single out-of-date follower as "unavailable" status in Raft fallback.
If we're the leader of a two node cluster, read-only TX's on a broken
network will continue to succeed until a read-write transaction happens,
because the local node is smart enough to know it still must be the leader;
this is because any new leader would require its vote.

Detect this case and fail the availability check if so.