Rebase MVCC transactions on commit to avoid having to keep multiple versions.
Instead of waiting until an MVCC transaction is committed to calculate whether
it has conflicts with any other transactions committed since it was created,
when other transactions are committed detect any conflicts with existing
transactions in realtime, marking conflicting transactions as failed and
"rebasing" non-conflicting transactions on the newly committed MVCC snapshot.

This means we only have to track one MVCC version (the current one), though
it also requires a new ability to safely swap out the KVStore underlying
a transaction's MutableView.