Fix a bug where reading a field within an @OnChange method returned a stale value.
@OnChange callbacks are invoked synchronously when fields are set. Previously,
if the invoked callback reads the field, and the field's value had previously
been cached, then the old cached value was being returned because the setter
method had not yet completed and updated it.

To avoid this problem, clear the 'cached' flag for the field just prior to
updating it, so any synchronous read will find the new value.

This fixes OnChangeNestedGetterTest.