Mark a Raft transaction's commit index valid as soon as we notice it.
Previously we were checking the validity of a transaction's commit index and
term each time we considered it for commit. Because followers may postpone
commits due to the leader lease timeout, it was possible for a transaction's
commit log entry to get applied to the state machine before it was ready to
commit, leading to a retry exception.

Once a transation's commit term and index are valid, they will always be,
so remember this after the first check and don't check again, eliminating
the possibility of a retry exception as described above.