Make Raft rethrow the same failure exception until commit/rollback.
Previously an asynchronously generated retry or other exception would be
thrown on the next transaction access, but any access after that would result
in a StaleTransactionException.

Unfortunately the Hibernate validator implementation swallows exceptions
thrown by property getters in PersistenceUtilHelper.isLoadedWithReference();
as a result, the transaction commit() ends up throwing StaleTransactionException
when it should instead be throwing RetryTransactionException.

To workaround this, change the Raft behavior so that the original exception is
thrown on every access (not just the first), until commit() or rollback() is
invoked. After that point, revert to throwing StaleTransactionException as before.