Fix bug where index queries failed on certain type hierarchies.
We were not properly handling the case where there is more than one maximally
narrow supertype of the target object type of a reference path (or, as a
special case, an index query target field). As a result, it was possible
for index queries and @OnChange method parameters specifying perfectly valid
generic types to be rejected as invalid if they used the "wrong" supertype.

In addition, we were not being strict enough in checking @OnChange parameter
types; if B and C extend A, and A has an @OnChange method for one of its fields,
the change event parameter needs to have generic type ? extends A, not just A.