Allow more flexibility for fields to change types across schema versions.
Now we allow any schema change as long as it doesn't result in conflicting
indexes, i.e., two different value types being encoded in indexes having the
same storage ID.

Now on object upgrade, if a field's type changes, the field must be reset.

Notable API changes:

  o Transaction.queryListElementIndex() and Transaction.queryMapValueIndex()
    now take the storage ID of the indexed sub-field rather than the parent.
  o New Transaction.addFieldChangeListener() accepts any kind of listener.

This is the first part of this change, affecting mainly the core API; the
high level API does not yet take full advantage of it.