Fix logic bug in KVDatabaseTest.
When a mutating transaction fails with a retry, not only do we not know
whether the transction actually committed, but also we don't know when
the transaction's mutations, if actually committed, will take effect. In
particular we can't assume that they don't take affect after the exception
is thrown and the transaction ends. In effect, a retry exception must be
treated like the transaction has left the party early and doesn't know what
happens after it leaves.

In this situation the logic that tracks the actual contents of the database
must therefore open and commit a new transaction that both mutates the data
and reads every key before it can assume that what it reads is truly accurate.