Raft bug fix and performance improvements.
Fix bug where Raft leaders did not respect their own required lease timeouts.

Improve Raft read-only transaction performance by capturing the commit index
as early as possible, i.e., as soon as the transaction is marked read-only,
instead of waiting until the transaction is committed. This still provides the
guarantee of linearizability, which only requires the read-only transaction's
view to be current as of some time between BEGIN and COMMIT, but eliminates
rebasing and conflict checking against mutating transactions whose BEGIN
time is strictly after the read-only transaction's BEGIN time.

In addition, for read-only transactions, do all conflict checking on the
follower instead of sending the reads to the leader for conflict checking.
Now we continue rebasing (and conflict checking) read-only transactions
on followers until the commit log entry, as reported by the leader, is
received. The benefit is that read-only transactions now only require a
constant amount of network traffic.

To do all this safely we must no longer allow read-only transactions to be
set back to read-write.