Improvements to how object copies are made.
Previously, only the first object copied could have its object ID remapped,
and this remapping only applied to the object ID, not any contained reference
field values.

Generalize the copy operation to allow a mapping of object IDs to be provided
and use it to fixup reference fields as we copy.