Optimize for multiple redundant invocations of AtomicArrayKVStore.snapshot().
After creating an immutable copy of the current mutation set (i.e., Writes)
for the snapshot() method, cache it so we don't repeat that work on subsequent
invocations if there have been no further mutations.

This is often the case when AtomicArrayKVStore is used underneatth an MVCC
database (e.g., ArrayKVDatabase or RaftKVDatabase) with a high ratio of read
to write transactions.