Close race window in MutableView when recording read ranges.
In MutableView, the reading of the underlying key/value store and the recording
of the corresponding read ranges must be done together while continuously
synchronized. Otherwise, a read of the key/value store could be missed by an
external observer who has locked the MutableView in order to take a "snapshot".