Improve RaftKVDatabase performance when under duress.
With heavy transaction load, and/or when there is a slow follower, the leader
can reach its limit on unapplied log entries.  As a result, it starts applying
them to the state machine before all followers have received them. This leads
to snapshot installs and more rejected transactions which cause additional
slowness.

The fix is to eliminate the requirement that log entries remain unapplied
in order to be sendable to followers or used for conflict detection. Now we
always apply immediately, but keep a copy of already-applied log entries
(up to a limit) for that purpose. To save memory, we discard the Writes
after application, and read from disk as needed for conflict detection.