Fix bug where snapshot tx could leak if commit() throws an error.
The semantics of commit() are such that if commit() throws an exception,
the caller is not required to invoke rollback(), whereas if an exception is
thrown during normal operation (e.g., retry exception when reading a key),
then the caller is still required to invoke rollback().

Therefore, if a snapshot transaction has a pending retry error put there
by another thread, and that error is detected by commit(), then we must
rollback() the transaction ourselves before throwing it to the caller.