Add ApplicationClassLoader and use it for Permazen model classes, etc.
Previously, to implement the "--classpath" in permazen-app, we were hacking the
system classloader using reflection. In JDK 9+, this no longer allowed (by default).

Instead, create a proper singleton ApplicationClassLoader class, whose parent is
the system (or context) class loader and which has a mutable list of URLs for
"--classpath" to modify, and update all Permazen code that loads user classes to
use this singleton loader.