Eliminate the MAX_MUTABLE_VIEW_DEPTH limit when creating transactions.
The laudable goal was to avoid having to traverse too many levels of unapplied
log entries when reading from the transaction's MutableView. However, because we
are always rebasing transactions as much as possible, such a transaction will
quickly be rebased up to the top-most log entry, so the effect will be at best
transient.

In addition, this limit, when exceeded, was triggering an assertion failure when
the transaction, which was expected to be fully rebased, was found not to be.