Refactor so most of Maven plugin logic is implemented in permazen-main.
This ensures that the schema-related logic that is applied during the build is
the same as the schema-related logic that will apply at runtime.

Also, this avoids class loader issues that occur with the Maven plugin. The
Maven plugin has its own class loader that doesn't include any of the project's
dependencies. We can (and do) create a new loader containing those dependencies
which delegates to the plugin loader as its parent, but when the plugin itself
depends on and links directly to any Permazen classes, those Permazen classes
will be loaded in the plugin's loader, and therefore be unable to "see" any
classes they (optionally) link to in the project dependencies. So for example
the Permazen optional built-in encoding for InternetAddress, which depends on
the optional jakarta.mail-api JAR, won't find InternetAddress.class, even if the
project includes jakarta.mail-api as a dependency.

Instead, perform all schema logic, including class loading of Permazen classes,
user model classes, and optional dependencies, within the context of a custom
class loader created for that purpose, and have the Maven plugin access the
schema functionality through the service interface SchemaUtility.